rename fiat plugin
