+++ src/constants/plugins.ts
@@ -9,7 +9,7 @@ import Feeds from '@assets/Plugins/Feeds.svg'
 import $1 from '@plugins/default-theme'
 import $2 from '@plugins/buy'
 import $3 from '@plugins/community'
-import $4 from '@plugins/fiat'
+import $4 from '@plugins/pay-by-card'
 import $5 from '@plugins/home'
 import $6 from '@plugins/join'
 import $7 from '@plugins/me'
