+++ src/plugins/pay-by-card/slot.astro
